Created Authorization service class for checking the permission to edit or delete an offer.

+++ ProfitApp/src/main/java/bsu/fpmi/profit/AdsServlet.java
@@ -6,6 +6,7 @@ import javax.servlet.ServletException;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 public class AdsServlet extends HttpServlet {
@@ -45,13 +46,35 @@ public class AdsServlet extends HttpServlet {
                     break;
                 }
                 case "edit": {
-                    response.getWriter().print((new Gson()).toJson(profitCRUD.editOffer(request.getParameter("id"),
-                            (new Gson()).fromJson(request.getReader().readLine(), Ad.class))));
+                    try{
+                        if(checkUserId(request)) {
+                            response.getWriter().print((new Gson()).toJson(profitCRUD.editOffer(request.getParameter("id"),
+                                    (new Gson()).fromJson(request.getReader().readLine(), Ad.class))));
+                        }
+                        else {
+                            String message = "You don't have permission to do that.";
+                            response.sendError(403, message);
+                        }
+                    }catch (OfferNotFoundException e){
+                        response.sendError(404, e.getMessage());
+                        e.printStackTrace();
+                    }
                     break;
                 }
                 case "review": {
-                    response.getWriter().print((new Gson()).toJson(profitCRUD.addReview(request.getParameter("id"),
-                            (new Gson()).fromJson(request.getReader().readLine(), Review.class))));
+                    try {
+                        if(!checkUserId(request)) {
+                            response.getWriter().print((new Gson()).toJson(profitCRUD.addReview(request.getParameter("id"),
+                                    (new Gson()).fromJson(request.getReader().readLine(), Review.class))));
+                        }
+                        else {
+                            String message = "You don't have permission to do that.";
+                            response.sendError(403, message);
+                        }
+                    } catch (OfferNotFoundException e){
+                        response.sendError(404, e.getMessage());
+                        e.printStackTrace();
+                    }
                     break;
                 }
             }
@@ -66,7 +89,28 @@ public class AdsServlet extends HttpServlet {
 
     @Override
     protected void doDelete(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        response.setContentType("json");
-        response.getWriter().print((new Gson()).toJson(profitCRUD.deleteAd(request.getParameter("id"))));
+        try {
+            if (checkUserId(request)) {
+                response.setContentType("json");
+                response.getWriter().print((new Gson()).toJson(profitCRUD.deleteAd(request.getParameter("id"))));
+            }
+            else {
+                String message = "You don't have permission to do that.";
+                response.sendError(403, message);
+            }
+        }catch (OfferNotFoundException e){
+            response.sendError(404, e.getMessage());
+            e.printStackTrace();
+        }
+    }
+    private boolean checkUserId(HttpServletRequest request) throws OfferNotFoundException {
+        HttpSession session = request.getSession(false);
+        if(session != null && session.getAttribute("USER") != null) {
+            String offerId = request.getParameter("id");
+            return AuthorizationService.isPostAvailable(Integer.parseInt(offerId),
+                    ((User) session.getAttribute("USER")).getUserId(),
+                    profitCRUD.getConnection());
+        }
+        return false;
     }
 }
